requirements.txt
absl-py==0.7.1
arrow==0.13.1
astor==0.7.1
astroid==2.2.5
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
future==0.17.1
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
importlib==1.0.4
isort==4.3.17
Jinja2==2.10.1
jinja2-time==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
numpy==1.16.3
pandas==0.24.2
pbr==5.2.0
poyo==0.4.2
protobuf==3.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.5
urllib3==1.24.2
Werkzeug==0.15.2
whichcraft==0.5.2
wrapt==1.11.1
